+++ include/DoocsProcessScalar.h
@@ -89,13 +89,13 @@ namespace ChimeraTK {
 
     void publishZeroMQ() { publishZMQ = true; }
 
-    void setMacroPulseNumberSource(boost::shared_ptr<ChimeraTK::NDRegisterAccessor<int>> macroPulseNumberSource) {
+    void setMacroPulseNumberSource(boost::shared_ptr<ChimeraTK::NDRegisterAccessor<int64_t>> macroPulseNumberSource) {
       _macroPulseNumberSource = macroPulseNumberSource;
     }
 
    protected:
     boost::shared_ptr<ChimeraTK::NDRegisterAccessor<T>> _processScalar;
-    boost::shared_ptr<ChimeraTK::NDRegisterAccessor<int>> _macroPulseNumberSource;
+    boost::shared_ptr<ChimeraTK::NDRegisterAccessor<int64_t>> _macroPulseNumberSource;
     bool publishZMQ{false};
   };
